13/12/19

+++ springtask/src/main/java/task/DAO/UserDAO.java
@@ -8,8 +8,8 @@ public interface UserDAO {
 
     public void addUser(UserDataSet uds);
     public void updateUser(UserDataSet uds);
-    public void removeUser(int id);
+    public void removeUser(long id);
     public UserDataSet getUserByLogin(String login);
-    public UserDataSet getUserById(Integer id);
+    public UserDataSet getUserById(long id);
     public List<UserDataSet> findAll();
 }
