+++ springtask/src/main/java/task/service/UserService.java
@@ -7,8 +7,8 @@ public interface UserService {
 
     public void addUser(UserDataSet uds);
     public void updateUser(UserDataSet uds);
-    public void removeUser(int id);
+    public void removeUser(long id);
     public UserDataSet getUserByLogin(String login);
-    public UserDataSet getUserById(Integer id);
+    public UserDataSet getUserById(long id);
     public List<UserDataSet> findAllUsers();
 }
